+++ src/main/java/com/svalero/music/rights/controller/WorkController.java
@@ -51,24 +51,11 @@ public class WorkController {
         workService.delete(id);
     }
 
-
     //FILTRADO
     @GetMapping("/works/by-musician/{id}")
     public List<Work> getByMusician(@PathVariable Long id) throws MusicianNotFoundException {
         List <Work> worksOfMusician = workService.findByMusician(id);
         return worksOfMusician;
-    }
-
-    @ExceptionHandler(WorkNotFoundException.class)
-    public ResponseEntity<ErrorResponse> handleException(WorkNotFoundException cnfe) {
-        ErrorResponse errorResponse = new ErrorResponse(404, "Not-found", "La obra no existe");
-        return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler(MusicianNotFoundException.class)
-    public ResponseEntity<ErrorResponse> handleException(MusicianNotFoundException cnfe) {
-        ErrorResponse errorResponse = new ErrorResponse(404, "Not-found", "El músico no existe");
-        return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
     }
 
 }
